algorithm-leetcode-jvm/src/main/java/org/muxun/algorithm/leetcode/java/array/LeetCode33.java
package org.muxun.algorithm.leetcode.java.array;

/**
 * 题号：33
 * <p>
 * 题目：<a href="https://leetcode.cn/problems/search-in-rotated-sorted-array">搜索旋转排序数组</a>
 * <p>
 * 详情：
 * <p>
 * 假设按照升序排序的数组在预先未知的某个点上进行了旋转。
 * <p>
 * (例如，数组 [0,1,2,4,5,6,7] 可能变为 [4,5,6,7,0,1,2])
 * <p>
 * 搜索一个给定的目标值，如果数组中存在这个目标值，则返回它的索引，否则返回 -1 。
 * <p>
 * 你可以假设数组中不存在重复的元素。
 * <p>
 * 你的算法时间复杂度必须是 O(log n) 级别。
 * @author 慕勋
 * @created 2020-02-17
 */
public class LeetCode33 {

    /**
     * 思路：二分查找
     * @param nums   给定的旋转数组
     * @param target 给定的目标值
     * @return 目标值在数组中的索引，不在数组中返回-1
     */
    public int search(int[] nums, int target) {
        int n = nums.length;
        if (n == 0) {
            return -1;
        }
        if (n == 1) {
            return nums[0] == target ? 0 : -1;
        }
        int left = 0;
        int right = n - 1;
        while (left <= right) {
            int mid = (left + right) / 2;
            if (nums[mid] == target) {
                return mid;
            }
            if (nums[0] <= nums[mid]) {
                if (nums[0] <= target && target < nums[mid]) {
                    right = mid - 1;
                } else {
                    left = mid + 1;
                }
            } else {
                if (nums[mid] < target && target <= nums[n - 1]) {
                    left = mid + 1;
                } else {
                    right = mid - 1;
                }
            }
        }
        return -1;
    }
}
